+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**20170317** 
+
+- [Copy-Module.ps1](Source\Scripts\PSRepository\Copy-Module.ps1). **[New]**
+
+
 **20170307** 
 
 - [Get-Github.ps1](Source\Scripts\Github\Get-Github.ps1) works with PowerShell v4.0
@@ -5,11 +10,11 @@
 **20161025** 
 
 - Bumped the version in [Get-Github.ps1](Source\Scripts\Github\Get-Github.ps1) to `0.2`.
-- Added support for tags in [Get-Github.ps1](Source\Scripts\Github\Get-Github.ps1) **[Update]**.
+- Added support for tags in [Get-Github.ps1](Source\Scripts\Github\Get-Github.ps1). **[Update]**
 
 # pre-release v0.1
 
 **20161024**
 
 - Initial commit to repository.
-- [Get-Github.ps1](Source\Scripts\Github\Get-Github.ps1) **[New]**.
+- [Get-Github.ps1](Source\Scripts\Github\Get-Github.ps1). **[New]**